chore: refactor some code to use int instead of int64

+++ cmd/api/entity/cart.go
@@ -7,8 +7,8 @@ import (
 // @Model
 type Cart struct {
 	BaseEntity
-	ProductId int64   `gorm:"type:bigint;not null;column:product_id" json:"product_id"`
-	UserId    int64   `gorm:"type:bigint;not null;column:user_id" json:"user_id"`
+	ProductId int   `gorm:"type:bigint;not null;column:product_id" json:"product_id"`
+	UserId    int   `gorm:"type:bigint;not null;column:user_id" json:"user_id"`
 	Quantity  int     `gorm:"type:integer;not null;column:quantity" json:"quantity"`
 	Product   Product `json:"product"`
 }
